Commit Final

+++ controllers/crudControllers.js
@@ -1,8 +1,9 @@
 const db = require('../config/db');
 const { validarContraseña } = require('../utils/validarContraseña');
+const bcrypt = require('bcrypt');
 
 // Función para agregar un usuario
-const addUser = (req, res) => {
+const addUser = async (req, res) => {
     const { usuario, fecha_nacimiento, email, contraseña, role_id } = req.body;
 
     // Validar la contraseña
@@ -10,39 +11,38 @@ const addUser = (req, res) => {
         return res.status(400).json({ mensaje: 'La contraseña no cumple con los requisitos de seguridad' });
     }
 
-    const query = 'INSERT INTO Usuarios (usuario, fecha_nacimiento, email, contraseña, role_id) VALUES (?, ?, ?, ?, ?)';
-    db.query(query, [usuario, fecha_nacimiento, email, contraseña, role_id], (err, result) => {
-        if (err) {
-            console.error('Error al agregar usuario:', err);
-            return res.status(500).json({ mensaje: 'Error del servidor al agregar usuario' });
-        } else {
-            return res.status(201).json({ mensaje: 'Usuario agregado exitosamente', userId: result.insertId });
-        }
-    });
-};
+    const hash = bcrypt.hashSync(contraseña, 10); // Asumiendo que estás usando bcrypt para el hash
 
+    try {
+        const [result] = await db.query('INSERT INTO Usuarios (usuario, fecha_nacimiento, email, contraseña, role_id) VALUES (?, ?, ?, ?, ?)', [usuario, fecha_nacimiento, email, hash, role_id]);
+        return res.status(201).json({ mensaje: 'Usuario agregado exitosamente', userId: result.insertId });
+    } catch (err) {
+        console.error('Error al agregar usuario:', err);
+        return res.status(500).json({ mensaje: 'Error del servidor al agregar usuario' });
+    }
+};
 // Función para eliminar un usuario
-const deleteUser = (req, res) => {
+const deleteUser = async (req, res) => {
     const userId = req.params.id;
-    const query = 'DELETE FROM Usuarios WHERE ID_Usuario = ?';
-    db.query(query, [userId], (err, result) => {
-        if (err) {
-            console.error('Error al eliminar usuario:', err);
-            return res.status(500).json({ mensaje: 'Error del servidor al eliminar usuario' });
-        } else if (result.affectedRows === 0) {
+
+    try {
+        const [result] = await db.query('DELETE FROM Usuarios WHERE ID_Usuario = ?', [userId]);
+
+        if (result.affectedRows === 0) {
             return res.status(404).json({ mensaje: 'Usuario no encontrado' });
-        } else {
-            return res.json({ mensaje: 'Usuario eliminado exitosamente' });
         }
-    });
+
+        return res.json({ mensaje: 'Usuario eliminado exitosamente' });
+    } catch (err) {
+        console.error('Error al eliminar usuario:', err);
+        return res.status(500).json({ mensaje: 'Error del servidor al eliminar usuario' });
+    }
 };
 
 // Función para actualizar un usuario
-const updateUser = (req, res) => {
+const updateUser = async (req, res) => {
     const userId = req.params.id;
     const { usuario, fecha_nacimiento, email, rol } = req.body;
-
-    console.log('Datos recibidos:', { userId, usuario, fecha_nacimiento, email, rol });
 
     if (!rol) {
         return res.status(400).json({ mensaje: 'El rol es requerido' });
@@ -51,20 +51,18 @@ const updateUser = (req, res) => {
     const query = 'UPDATE Usuarios SET usuario = ?, fecha_nacimiento = ?, email = ?, role_id = ? WHERE ID_Usuario = ?';
     const params = [usuario, fecha_nacimiento, email, rol, userId];
 
-    console.log('Consulta de actualización:', query);
-    console.log('Parámetros de actualización:', params);
+    try {
+        const [result] = await db.query(query, params);
 
-    db.query(query, params, (err, result) => {
-        if (err) {
-            console.error('Error al actualizar usuario:', err);
-            return res.status(500).json({ mensaje: 'Error del servidor al actualizar usuario' });
-        } else if (result.affectedRows === 0) {
+        if (result.affectedRows === 0) {
             return res.status(404).json({ mensaje: 'Usuario no encontrado' });
-        } else {
-            console.log('Resultado de la actualización:', result);
-            return res.json({ mensaje: 'Usuario actualizado exitosamente' });
         }
-    });
+
+        return res.json({ mensaje: 'Usuario actualizado exitosamente' });
+    } catch (err) {
+        console.error('Error al actualizar usuario:', err);
+        return res.status(500).json({ mensaje: 'Error del servidor al actualizar usuario' });
+    }
 };
 
 module.exports = {
